SIMPLY-3621 Refactor BookButtonsView with BookDetailButtonsView
BookDetailButtonsView is a superset of BookButtonsView, except for the handling of the "clock" icon, which is taken from BookButtonsView.

NYPLBookNormalCell (a non-Auto Layout view using it) and NYPLRoundedButton needed adjusting since NYPLBookDetailButtonsView does use Auto Layout.
