+++ Simplified/NYPLBookCellDelegate.h
@@ -1,7 +1,6 @@
-#import "NYPLBookNormalCell.h"
 #import "NYPLBookDownloadFailedCell.h"
 #import "NYPLBookDownloadingCell.h"
-#import "NYPLBookButtonsView.h"
+#import "NYPLBookDetailButtonsView.h"
 
 /* This class implements a shared delegate that performs all of its duties via the shared registry,
 shared cover registry, shared download center, et cetera. */
