+++ Simplified/NYPLBookContentType.swift
